Config mongoose to use the new Server Discovery and Monitoring
Without this config, by default, mongoose.connect() will print out the below warning:
```
DeprecationWarning: current Server Discovery and Monitoring engine is deprecated, and will be removed in a future version. To use the new Server Discover and Monitoring engine, pass option { useUnifiedTopology: true } to the MongoClient constructor.

```

The application needs to opt-in using the new topology engine, using this change, to suppress the warning.